+++ TeleRequestHandler/repo_service/repo_register_service.go
@@ -5,8 +5,8 @@ import (
 )
 
 type RepoRegisterService interface {
-	AddRepo(chatId int, dto Common.RepoDTO) error
-	GetReposByChat(chatId int) ([]Common.RepoDTO, error)
-	GetReposByTag(chatId int, tag string) ([]Common.RepoDTO, error)
-	DeleteRepo(chatId int, link string) error
+	AddRepo(chatId int64, dto Common.RepoDTO) error
+	GetReposByChat(chatId int64) ([]Common.RepoDTO, error)
+	GetReposByTag(chatId int64, tag string) ([]Common.RepoDTO, error)
+	DeleteRepo(chatId int64, link string) error
 }
